Reduced complexity, refactored
